+++ biomarker/docker_push_tags.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-docker push prismcmap/biomarker-module:v0.1.1
-docker push prismcmap/biomarker-module:latest
+docker push prismcmap/biomarker:testing
 
 # trigger