+++ drc/docker_push_tags.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-docker push prismcmap/drc-module:v0.0.1
-docker push prismcmap/drc-module:latest
+#docker push prismcmap/drc-module:v0.0.1
+docker push prismcmap/drc:production
 
 # trigger