+++ stack/stack.py
@@ -69,6 +69,11 @@ def main(args):
             'search_pattern': '*_cell_info.txt',
             'type': 'metadata'
         },
+        'EPS_QC_TABLE': {
+            'search_pattern': '*QC_TABLE*.csv',
+            'type': 'report',
+            'format': 'csv'
+        },
         'QC_TABLE': {
             'search_pattern': '*QC_TABLE*.csv',
             'type': 'report',
@@ -89,9 +94,9 @@ def main(args):
             'search_pattern': '*_LEVEL3_LMFI*.csv',
             'type': 'csv_data'
         },
-        'LEVEL3_NORMALIZED': {
-                    'search_pattern': '*_LEVEL3_NORMALIZED*.csv',
-                    'type': 'csv_data'
+        'LEVEL3_NORMALIZED_COUNTS': {
+            'search_pattern': '*_LEVEL3_NORMALIZED_COUNTS*.csv',
+            'type': 'csv_data'
                 },
         'LEVEL4_LFC': {
             'search_pattern': '*_LEVEL4_LFC*.csv',